Refactore wallet structure (part 1)
- Rename UserStakeScreen -> StakeView and remove all unneeded content
- Fix: `addPhrase` to return Promise properly
- Fix: `preview.txs` to use latest contexts
- Extract content of `UnlockView` into `UnlockForm` and add a story for
it
