+++ .storybook/preview.tsx
@@ -6,7 +6,7 @@ import { configure } from '@storybook/react'
 import { withThemes } from '@react-theming/storybook-addon'
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
-import { ConnectionProvider } from '../src/contexts/ConnectionContext'
+import { AppProvider } from '../src/contexts/AppContext'
 import { ThemeProvider } from '../src/contexts/ThemeContext'
 import { I18nProvider } from '../src/contexts/I18nContext'
 import { AppWrapper } from '../src/App.style'
@@ -18,14 +18,14 @@ const darkTheme = { name: 'Dark', ...themes.dark }
 addDecorator(withKnobs)
 
 const providerFn = ({ theme, children }) => (
-  <ConnectionProvider>
+  <AppProvider>
     <I18nProvider>
       <ThemeProvider theme={theme}>
         <GlobalStyle />
         <AppWrapper>{children}</AppWrapper>
       </ThemeProvider>
     </I18nProvider>
-  </ConnectionProvider>
+  </AppProvider>
 )
 
 addDecorator(withThemes(null, [lightTheme, darkTheme], { providerFn }))
