+++ src/views/wallet/StakesView.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const StakesView: React.FC = (): JSX.Element => <h1>StakesView</h1>
+
+export default StakesView
